+++ card_engine/src/legality.rs
@@ -1,0 +1,86 @@
+// Legalities pack into a u64: 2 bits per format, positions handed out append-only
+// by a global registry the first time a format name appears in loaded data, so
+// bit assignments stay stable across reloads and engine instances. A format the
+// card's JSONB omits reads as not_legal. 32 formats fit; Scryfall ships 22.
+
+use std::collections::HashMap;
+use std::sync::{OnceLock, RwLock};
+use pyo3::prelude::*;
+use pyo3::types::PyDict;
+use rkyv::Archived;
+
+const LEGALITY_NOT_LEGAL: u64 = 0;
+pub(crate) const LEGALITY_LEGAL: u64 = 1;
+pub(crate) const LEGALITY_RESTRICTED: u64 = 2;
+pub(crate) const LEGALITY_BANNED: u64 = 3;
+const MAX_FORMATS: usize = 32;
+
+static FORMAT_SHIFTS: OnceLock<RwLock<HashMap<String, u8>>> = OnceLock::new();
+
+pub(crate) fn format_shifts() -> &'static RwLock<HashMap<String, u8>> {
+    FORMAT_SHIFTS.get_or_init(|| RwLock::new(HashMap::new()))
+}
+
+/// Bit shift for a format already seen in loaded data; None matches nothing.
+pub(crate) fn format_shift(format: &str) -> Option<u8> {
+    format_shifts().read().ok()?.get(format).copied()
+}
+
+/// Bit shift for a format, assigning the next free slot if unseen (reload path).
+pub(crate) fn format_shift_or_assign(format: &str) -> Option<u8> {
+    if let Some(shift) = format_shift(format) {
+        return Some(shift);
+    }
+    let mut shifts = format_shifts().write().ok()?;
+    if let Some(&shift) = shifts.get(format) {
+        return Some(shift); // assigned while we waited for the write lock
+    }
+    if shifts.len() >= MAX_FORMATS {
+        return None;
+    }
+    let shift = (shifts.len() * 2) as u8;
+    shifts.insert(format.to_string(), shift);
+    Some(shift)
+}
+
+fn legality_code(status: &str) -> u64 {
+    match status {
+        "legal"      => LEGALITY_LEGAL,
+        "restricted" => LEGALITY_RESTRICTED,
+        "banned"     => LEGALITY_BANNED,
+        _            => LEGALITY_NOT_LEGAL,
+    }
+}
+
+pub(crate) fn jsonb_obj_to_legality_bits(d: &Bound<PyDict>, key: &str) -> u64 {
+    d.get_item(key)
+        .ok()
+        .flatten()
+        .and_then(|v| {
+            v.cast::<PyDict>().ok().map(|m| {
+                m.iter()
+                    .filter_map(|(k, v)| {
+                        let format = k.extract::<String>().ok()?;
+                        let status = v.extract::<String>().ok()?;
+                        let shift = format_shift_or_assign(&format)?;
+                        Some(legality_code(&status) << shift)
+                    })
+                    .fold(0u64, |bits, b| bits | b)
+            })
+        })
+        .unwrap_or_default()
+}
+
+/// Adopt the archive's format→shift assignments into this process's registry.
+/// Cheap no-op (one read lock) once the registry has caught up.
+pub(crate) fn sync_format_shifts(archived: &Archived<HashMap<String, u8>>) {
+    let behind = format_shifts().read().map(|m| m.len() < archived.len()).unwrap_or(false);
+    if !behind {
+        return;
+    }
+    if let Ok(mut shifts) = format_shifts().write() {
+        for (format, shift) in archived.iter() {
+            shifts.insert(format.as_str().to_string(), *shift);
+        }
+    }
+}
