+++ card_engine/src/lib.rs
@@ -1,13 +1,12 @@
 use pyo3::prelude::*;
 use pyo3::types::{PyDate, PyDateAccess, PyDict, PyList, PyTuple};
-use regex::Regex;
 use rkyv::{Archive, Archived, Deserialize, Serialize};
 use memmap2::Mmap;
 use serde_json::Value;
 use std::collections::{HashMap, HashSet};
 use std::io::Write as IoWrite;
 use std::path::PathBuf;
-use std::sync::{Arc, Mutex, OnceLock, RwLock};
+use std::sync::{Arc, Mutex};
 use std::os::unix::io::AsRawFd;
 use std::os::unix::fs::MetadataExt;
 
@@ -16,160 +15,12 @@ use std::os::unix::fs::MetadataExt;
 // a breakdown of reload(): see docs/issues/engine-store-size-reduction.md step 0.
 
 #[cfg(feature = "alloc-counter")]
-mod alloc_stats {
-    use std::alloc::{GlobalAlloc, Layout, System};
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    pub static LIVE: AtomicUsize = AtomicUsize::new(0);
-    pub static PEAK: AtomicUsize = AtomicUsize::new(0);
-    pub static ALLOCS: AtomicUsize = AtomicUsize::new(0); // currently-live allocation count
-
-    // Snapshots recorded by the most recent reload()
-    pub static RELOAD_LIVE_BEFORE: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_LIVE_AFTER_CARDS: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_ALLOCS_AFTER_CARDS: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_LIVE_AFTER_INDEXES: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_ALLOCS_AFTER_INDEXES: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_CARDS_RKYV: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_INDEXES_RKYV: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_STRINGS_RKYV: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_ARCHIVE: AtomicUsize = AtomicUsize::new(0);
-    pub static RELOAD_PEAK: AtomicUsize = AtomicUsize::new(0);
-
-    pub fn live() -> usize { LIVE.load(Ordering::Relaxed) }
-    pub fn allocs() -> usize { ALLOCS.load(Ordering::Relaxed) }
-    pub fn peak() -> usize { PEAK.load(Ordering::Relaxed) }
-
-    pub fn reset_peak() {
-        RELOAD_LIVE_BEFORE.store(live(), Ordering::Relaxed);
-        PEAK.store(live(), Ordering::Relaxed);
-    }
-
-    pub fn record_reload(
-        after_cards: (usize, usize),
-        after_indexes: (usize, usize),
-        component_bytes: (usize, usize, usize),
-        archive: usize,
-        peak: usize, // caller snapshots before diagnostics inflate the high-water mark
-    ) {
-        RELOAD_LIVE_AFTER_CARDS.store(after_cards.0, Ordering::Relaxed);
-        RELOAD_ALLOCS_AFTER_CARDS.store(after_cards.1, Ordering::Relaxed);
-        RELOAD_LIVE_AFTER_INDEXES.store(after_indexes.0, Ordering::Relaxed);
-        RELOAD_ALLOCS_AFTER_INDEXES.store(after_indexes.1, Ordering::Relaxed);
-        RELOAD_CARDS_RKYV.store(component_bytes.0, Ordering::Relaxed);
-        RELOAD_INDEXES_RKYV.store(component_bytes.1, Ordering::Relaxed);
-        RELOAD_STRINGS_RKYV.store(component_bytes.2, Ordering::Relaxed);
-        RELOAD_ARCHIVE.store(archive, Ordering::Relaxed);
-        RELOAD_PEAK.store(peak, Ordering::Relaxed);
-    }
-
-    pub struct CountingAlloc;
-
-    impl CountingAlloc {
-        fn on_alloc(size: usize) {
-            let live = LIVE.fetch_add(size, Ordering::Relaxed) + size;
-            ALLOCS.fetch_add(1, Ordering::Relaxed);
-            PEAK.fetch_max(live, Ordering::Relaxed);
-        }
-        fn on_dealloc(size: usize) {
-            LIVE.fetch_sub(size, Ordering::Relaxed);
-            ALLOCS.fetch_sub(1, Ordering::Relaxed);
-        }
-    }
-
-    unsafe impl GlobalAlloc for CountingAlloc {
-        unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-            let p = unsafe { System.alloc(layout) };
-            if !p.is_null() { Self::on_alloc(layout.size()); }
-            p
-        }
-        unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
-            unsafe { System.dealloc(ptr, layout) };
-            Self::on_dealloc(layout.size());
-        }
-        unsafe fn realloc(&self, ptr: *mut u8, layout: Layout, new_size: usize) -> *mut u8 {
-            let p = unsafe { System.realloc(ptr, layout, new_size) };
-            if !p.is_null() {
-                LIVE.fetch_sub(layout.size(), Ordering::Relaxed);
-                let live = LIVE.fetch_add(new_size, Ordering::Relaxed) + new_size;
-                PEAK.fetch_max(live, Ordering::Relaxed);
-            }
-            p
-        }
-        unsafe fn alloc_zeroed(&self, layout: Layout) -> *mut u8 {
-            let p = unsafe { System.alloc_zeroed(layout) };
-            if !p.is_null() { Self::on_alloc(layout.size()); }
-            p
-        }
-    }
-
-    #[global_allocator]
-    static COUNTING_ALLOC: CountingAlloc = CountingAlloc;
-}
+mod alloc_stats;
 
 // ─── Inline string (no heap allocation) ──────────────────────────────────────
 
-// repr(C) guarantees a stable layout ([u8; N] then u8, align 1, no padding)
-// across compiler versions, as the Portable impl below requires.
-#[repr(C)]
-#[derive(Clone, Copy)]
-struct InlineStr<const N: usize> {
-    bytes: [u8; N],
-    len: u8,
-}
-
-// Safety: InlineStr<N> is repr(C) with only align-1 fields — a stable, fully
-// initialized, padding-free layout — and carries no internal references, so it
-// is safe to treat as a flat, relocatable value in an rkyv archive.
-unsafe impl<const N: usize> rkyv::Portable for InlineStr<N> {}
-
-impl<const N: usize> Archive for InlineStr<N> {
-    type Archived = InlineStr<N>;
-    type Resolver = ();
-    fn resolve(&self, _: (), out: rkyv::Place<InlineStr<N>>) {
-        // Safety: InlineStr<N> is Copy and Portable; writing it verbatim is correct.
-        unsafe { out.ptr().write(*self); }
-    }
-}
-
-impl<const N: usize, S: rkyv::rancor::Fallible + ?Sized> Serialize<S> for InlineStr<N> {
-    fn serialize(&self, _serializer: &mut S) -> Result<(), S::Error> { Ok(()) }
-}
-
-impl<const N: usize, D: rkyv::rancor::Fallible + ?Sized> Deserialize<InlineStr<N>, D> for InlineStr<N> {
-    fn deserialize(&self, _: &mut D) -> Result<InlineStr<N>, D::Error> { Ok(*self) }
-}
-
-// Deliberately permissive: this impl trusts the data rather than validating it
-// (a real check would verify len <= N and UTF-8, since as_str() converts
-// unchecked). It exists only to satisfy the derived CheckBytes bounds on the
-// archived containers; validation is never the engine's safety boundary — the
-// archive is trusted by construction (see the access_unchecked justification
-// in QueryEngine::query()), so checked access is not relied on for soundness.
-unsafe impl<const N: usize, C: rkyv::rancor::Fallible + ?Sized> rkyv::bytecheck::CheckBytes<C> for InlineStr<N> {
-    unsafe fn check_bytes(
-        _value: *const Self,
-        _context: &mut C,
-    ) -> Result<(), C::Error> {
-        Ok(())
-    }
-}
-
-impl<const N: usize> InlineStr<N> {
-    fn from_str(s: &str) -> Self {
-        let max = s.len().min(N);
-        // Walk back from max to ensure we don't split a multi-byte char.
-        let len = (0..=max).rev().find(|&i| s.is_char_boundary(i)).unwrap_or(0);
-        let mut bytes = [0u8; N];
-        bytes[..len].copy_from_slice(&s.as_bytes()[..len]);
-        InlineStr { bytes, len: len as u8 }
-    }
-
-    #[inline]
-    fn as_str(&self) -> &str {
-        unsafe { std::str::from_utf8_unchecked(&self.bytes[..self.len as usize]) }
-    }
-}
+mod inline_str;
+use inline_str::InlineStr;
 
 // ─── Card type bits (u16) ─────────────────────────────────────────────────────
 
@@ -188,7 +39,7 @@ const TYPE_SNOW:         u16 = 1 << 11;
 const TYPE_SORCERY:      u16 = 1 << 12;
 const TYPE_WORLD:        u16 = 1 << 13;
 
-fn card_type_str_to_bit(s: &str) -> u16 {
+pub(crate) fn card_type_str_to_bit(s: &str) -> u16 {
     match s {
         "Artifact"     => TYPE_ARTIFACT,
         "Basic"        => TYPE_BASIC,
@@ -227,7 +78,7 @@ fn color_to_bit(c: &str) -> u8 {
     }
 }
 
-fn color_list_to_mask(colors: &[&str]) -> u8 {
+pub(crate) fn color_list_to_mask(colors: &[&str]) -> u8 {
     colors.iter().fold(0u8, |acc, c| acc | color_to_bit(c))
 }
 
@@ -235,11 +86,11 @@ fn color_list_to_mask(colors: &[&str]) -> u8 {
 
 /// Symbols that contribute to devotion, matching calculate_devotion() in SQL
 /// (which counts only WUBRGC characters of the mana cost string).
-fn is_devotion_sym(s: &str) -> bool {
+pub(crate) fn is_devotion_sym(s: &str) -> bool {
     s.len() == 1 && "WUBRGC".contains(s)
 }
 
-fn mana_pip_counts(s: &str) -> HashMap<String, u8> {
+pub(crate) fn mana_pip_counts(s: &str) -> HashMap<String, u8> {
     let mut pips: HashMap<String, u8> = HashMap::new();
     let upper = s.to_uppercase();
     let mut in_brace = false;
@@ -261,7 +112,7 @@ fn mana_pip_counts(s: &str) -> HashMap<String, u8> {
     pips
 }
 
-fn mana_cmc(s: &str) -> f32 {
+pub(crate) fn mana_cmc(s: &str) -> f32 {
     let upper = s.to_uppercase();
     let mut cmc = 0.0f32;
     let mut in_brace = false;
@@ -357,15 +208,15 @@ struct Card {
 }
 
 // Type alias for the archived (mmap-backed) card type
-type ACard = Archived<Card>;
+pub(crate) type ACard = Archived<Card>;
 // Archived string table (CardData.strings)
-type AStrings = Archived<Vec<String>>;
+pub(crate) type AStrings = Archived<Vec<String>>;
 
 /// Sentinel id for absent optional strings (a card never has 4 billion distinct strings).
 const NONE_STR: u32 = u32::MAX;
 
 /// Resolve an interned id against the archived string table; None for absent.
-fn str_at(strings: &AStrings, id: u32) -> Option<&str> {
+pub(crate) fn str_at(strings: &AStrings, id: u32) -> Option<&str> {
     if id == NONE_STR { None } else { Some(strings[id as usize].as_str()) }
 }
 
@@ -519,73 +370,9 @@ fn jsonb_obj_to_hashset(d: &Bound<PyDict>, key: &str) -> HashSet<String> {
 }
 
 // ─── Format legality bitmap ──────────────────────────────────────────────────
-//
-// Legalities pack into a u64: 2 bits per format, positions handed out append-only
-// by a global registry the first time a format name appears in loaded data, so
-// bit assignments stay stable across reloads and engine instances. A format the
-// card's JSONB omits reads as not_legal. 32 formats fit; Scryfall ships 22.
 
-const LEGALITY_NOT_LEGAL: u64 = 0;
-const LEGALITY_LEGAL: u64 = 1;
-const LEGALITY_RESTRICTED: u64 = 2;
-const LEGALITY_BANNED: u64 = 3;
-const MAX_FORMATS: usize = 32;
-
-static FORMAT_SHIFTS: OnceLock<RwLock<HashMap<String, u8>>> = OnceLock::new();
-
-fn format_shifts() -> &'static RwLock<HashMap<String, u8>> {
-    FORMAT_SHIFTS.get_or_init(|| RwLock::new(HashMap::new()))
-}
-
-/// Bit shift for a format already seen in loaded data; None matches nothing.
-fn format_shift(format: &str) -> Option<u8> {
-    format_shifts().read().ok()?.get(format).copied()
-}
-
-/// Bit shift for a format, assigning the next free slot if unseen (reload path).
-fn format_shift_or_assign(format: &str) -> Option<u8> {
-    if let Some(shift) = format_shift(format) {
-        return Some(shift);
-    }
-    let mut shifts = format_shifts().write().ok()?;
-    if let Some(&shift) = shifts.get(format) {
-        return Some(shift); // assigned while we waited for the write lock
-    }
-    if shifts.len() >= MAX_FORMATS {
-        return None;
-    }
-    let shift = (shifts.len() * 2) as u8;
-    shifts.insert(format.to_string(), shift);
-    Some(shift)
-}
-
-fn legality_code(status: &str) -> u64 {
-    match status {
-        "legal"      => LEGALITY_LEGAL,
-        "restricted" => LEGALITY_RESTRICTED,
-        "banned"     => LEGALITY_BANNED,
-        _            => LEGALITY_NOT_LEGAL,
-    }
-}
-
-fn jsonb_obj_to_legality_bits(d: &Bound<PyDict>, key: &str) -> u64 {
-    d.get_item(key)
-        .ok()
-        .flatten()
-        .and_then(|v| {
-            v.cast::<PyDict>().ok().map(|m| {
-                m.iter()
-                    .filter_map(|(k, v)| {
-                        let format = k.extract::<String>().ok()?;
-                        let status = v.extract::<String>().ok()?;
-                        let shift = format_shift_or_assign(&format)?;
-                        Some(legality_code(&status) << shift)
-                    })
-                    .fold(0u64, |bits, b| bits | b)
-            })
-        })
-        .unwrap_or_default()
-}
+mod legality;
+use legality::*;
 
 fn mana_cost_from_pydict(d: &Bound<PyDict>, cmc_val: Option<f32>) -> ManaCost {
     let pips: HashMap<String, u8> = d
@@ -694,894 +481,10 @@ fn card_from_pydict(d: &Bound<PyDict>, it: &mut Interner) -> Card {
     }
 }
 
-// ─── Filter expression ────────────────────────────────────────────────────────
+// ─── Filter expression & builder ─────────────────────────────────────────────
 
-#[derive(Clone, Copy, PartialEq)]
-enum CmpOp {
-    Eq,
-    Ne,
-    Lt,
-    Le,
-    Gt,
-    Ge,
-}
-
-#[derive(Clone, Copy)]
-enum ArithOp {
-    Add,
-    Sub,
-    Mul,
-    Div,
-}
-
-#[derive(Clone, Copy)]
-enum NumField {
-    Cmc,
-    Power,
-    Toughness,
-    Loyalty,
-    RarityInt,
-    CollectorNumberInt,
-    EdhrEc,
-    PriceUsd,
-    PriceEur,
-    PriceTix,
-    PreferScore,
-}
-
-impl NumField {
-    fn is_card_level(self) -> bool {
-        matches!(self, NumField::Cmc | NumField::Power | NumField::Toughness | NumField::Loyalty | NumField::EdhrEc)
-    }
-}
-
-fn attr_to_num_field(attr: &str) -> Option<NumField> {
-    match attr {
-        "cmc"                  => Some(NumField::Cmc),
-        "creature_power"       => Some(NumField::Power),
-        "creature_toughness"   => Some(NumField::Toughness),
-        "planeswalker_loyalty" => Some(NumField::Loyalty),
-        "card_rarity_int"      => Some(NumField::RarityInt),
-        "collector_number_int" => Some(NumField::CollectorNumberInt),
-        "edhrec_rank"          => Some(NumField::EdhrEc),
-        "price_usd"            => Some(NumField::PriceUsd),
-        "price_eur"            => Some(NumField::PriceEur),
-        "price_tix"            => Some(NumField::PriceTix),
-        "prefer_score"         => Some(NumField::PreferScore),
-        _ => None,
-    }
-}
-
-fn card_num(card: &ACard, f: NumField) -> Option<f32> {
-    match f {
-        NumField::Cmc                => card.cmc.as_ref().map(|v| u8::from(*v) as f32),
-        NumField::Power              => card.creature_power.as_ref().map(|v| i8::from(*v) as f32),
-        NumField::Toughness          => card.creature_toughness.as_ref().map(|v| i8::from(*v) as f32),
-        NumField::Loyalty            => card.planeswalker_loyalty.as_ref().map(|v| u8::from(*v) as f32),
-        NumField::RarityInt          => card.card_rarity_int.as_ref().map(|v| u8::from(*v) as f32),
-        NumField::CollectorNumberInt => card.collector_number_int.as_ref().map(|v| u16::from(*v) as f32),
-        NumField::EdhrEc             => card.edhrec_rank.as_ref().map(|v| u32::from(*v) as f32),
-        NumField::PriceUsd           => card.price_usd.as_ref().map(|v| f32::from(*v)),
-        NumField::PriceEur           => card.price_eur.as_ref().map(|v| f32::from(*v)),
-        NumField::PriceTix           => card.price_tix.as_ref().map(|v| f32::from(*v)),
-        NumField::PreferScore        => card.prefer_score.as_ref().map(|v| f32::from(*v)),
-    }
-}
-
-enum NumExpr {
-    Const(f64),
-    Field(NumField),
-    Arith(Box<NumExpr>, ArithOp, Box<NumExpr>),
-}
-
-impl NumExpr {
-    fn eval(&self, card: &ACard) -> Option<f64> {
-        match self {
-            NumExpr::Const(v) => Some(*v),
-            NumExpr::Field(f) => card_num(card, *f).map(|v| v as f64),
-            NumExpr::Arith(lhs, op, rhs) => {
-                let l = lhs.eval(card)?;
-                let r = rhs.eval(card)?;
-                Some(match op {
-                    ArithOp::Add => l + r,
-                    ArithOp::Sub => l - r,
-                    ArithOp::Mul => l * r,
-                    ArithOp::Div => {
-                        if r == 0.0 { return None; }
-                        l / r
-                    }
-                })
-            }
-        }
-    }
-
-    fn is_card_level(&self) -> bool {
-        match self {
-            NumExpr::Const(_) => true,
-            NumExpr::Field(f) => f.is_card_level(),
-            NumExpr::Arith(lhs, _, rhs) => lhs.is_card_level() && rhs.is_card_level(),
-        }
-    }
-}
-
-fn cmp(op: CmpOp, a: f64, b: f64) -> bool {
-    match op {
-        CmpOp::Eq => a == b,
-        CmpOp::Ne => a != b,
-        CmpOp::Lt => a < b,
-        CmpOp::Le => a <= b,
-        CmpOp::Gt => a > b,
-        CmpOp::Ge => a >= b,
-    }
-}
-
-#[derive(Clone, Copy)]
-enum ColorField {
-    Colors,
-    ColorIdentity,
-    ProducedMana,
-}
-
-fn card_colors(card: &ACard, f: ColorField) -> u8 {
-    match f {
-        ColorField::Colors        => u8::from(card.card_colors),
-        ColorField::ColorIdentity => u8::from(card.card_color_identity),
-        ColorField::ProducedMana  => u8::from(card.produced_mana),
-    }
-}
-
-#[derive(Clone, Copy)]
-enum CollField {
-    Subtypes,
-    Keywords,
-    OracleTags,
-    ArtTags,
-    IsTags,
-    FrameData,
-}
-
-impl CollField {
-    fn is_card_level(self) -> bool {
-        !matches!(self, CollField::ArtTags | CollField::FrameData | CollField::IsTags)
-    }
-}
-
-fn card_collection<'a>(card: &'a ACard, f: CollField) -> CollRef<'a> {
-    match f {
-        CollField::Subtypes   => CollRef::List(&card.card_subtypes),
-        CollField::Keywords   => CollRef::Set(&card.card_keywords),
-        CollField::OracleTags => CollRef::Set(&card.card_oracle_tags),
-        CollField::ArtTags    => CollRef::Set(&card.card_art_tags),
-        CollField::IsTags     => CollRef::Set(&card.card_is_tags),
-        CollField::FrameData  => CollRef::Set(&card.card_frame_data),
-    }
-}
-
-enum CollRef<'a> {
-    List(&'a rkyv::vec::ArchivedVec<rkyv::string::ArchivedString>),
-    Set(&'a rkyv::collections::swiss_table::ArchivedHashSet<rkyv::string::ArchivedString>),
-}
-
-impl CollRef<'_> {
-    fn contains(&self, v: &str) -> bool {
-        match self {
-            CollRef::List(l) => l.iter().any(|s| s.as_str() == v),
-            CollRef::Set(s)  => s.contains(v),
-        }
-    }
-    fn len(&self) -> usize {
-        match self {
-            CollRef::List(l) => l.len(),
-            CollRef::Set(s)  => s.len(),
-        }
-    }
-    fn all_equal(&self, v: &str) -> bool {
-        match self {
-            CollRef::List(l) => l.iter().all(|s| s.as_str() == v),
-            CollRef::Set(s)  => s.iter().all(|s| s.as_str() == v),
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq)]
-enum TextSearchField {
-    NameLower,
-    OracleTextLower,
-    FlavorTextLower,
-    ArtistLower,
-}
-
-impl TextSearchField {
-    fn is_card_level(self) -> bool {
-        matches!(self, TextSearchField::NameLower | TextSearchField::OracleTextLower)
-    }
-}
-
-fn text_search_field_value<'a>(card: &'a ACard, strings: &'a AStrings, field: TextSearchField) -> Option<&'a str> {
-    match field {
-        TextSearchField::NameLower       => Some(card.card_name_lower.as_str()),
-        TextSearchField::OracleTextLower => str_at(strings, u32::from(card.oracle_text_lower_id)),
-        TextSearchField::FlavorTextLower => str_at(strings, u32::from(card.flavor_text_lower_id)),
-        TextSearchField::ArtistLower     => str_at(strings, u32::from(card.card_artist_lower_id)),
-    }
-}
-
-/// Enum that replaces fn-pointer fields in TextExact / TextRegex.
-/// Function pointers cannot be parameterized over &Card vs &ACard, so enum
-/// dispatch is used instead.
-#[derive(Clone, Copy)]
-enum TextField {
-    NameLower,
-    OracleTextLower,
-    FlavorTextLower,
-    ArtistLower,
-    SetCode,
-    Layout,
-    Border,
-    Watermark,
-    CollectorNumber,
-}
-
-impl TextField {
-    fn is_card_level(self) -> bool {
-        matches!(self, TextField::NameLower | TextField::OracleTextLower)
-    }
-}
-
-fn text_field_value<'a>(card: &'a ACard, strings: &'a AStrings, field: TextField) -> Option<&'a str> {
-    match field {
-        TextField::NameLower       => Some(card.card_name_lower.as_str()),
-        TextField::OracleTextLower => str_at(strings, u32::from(card.oracle_text_lower_id)),
-        TextField::FlavorTextLower => str_at(strings, u32::from(card.flavor_text_lower_id)),
-        TextField::ArtistLower     => str_at(strings, u32::from(card.card_artist_lower_id)),
-        TextField::SetCode         => Some(card.card_set_code.as_str()),
-        TextField::Layout          => str_at(strings, u32::from(card.card_layout_id)),
-        TextField::Border          => str_at(strings, u32::from(card.card_border_id)),
-        TextField::Watermark       => str_at(strings, u32::from(card.card_watermark_id)),
-        TextField::CollectorNumber => str_at(strings, u32::from(card.collector_number_id)),
-    }
-}
-
-enum FilterExpr {
-    True,
-    And(Vec<FilterExpr>),
-    Or(Vec<FilterExpr>),
-    Not(Box<FilterExpr>),
-    ExactName(String),
-
-    NumericCmp {
-        lhs: NumExpr,
-        op: CmpOp,
-        rhs: NumExpr,
-    },
-
-    TextContains {
-        field: TextSearchField,
-        word: String,
-    },
-    TextExact {
-        field: TextField,
-        op: CmpOp,
-        value: String,
-    },
-    TextRegex {
-        field: TextField,
-        regex: Regex,
-    },
-
-    ColorCmp {
-        field: ColorField,
-        op: CmpOp,
-        mask: u8,
-    },
-
-    TypeCmp {
-        mask: u16,
-        op: CmpOp,
-    },
-
-    CollectionCmp {
-        field: CollField,
-        op: CmpOp,
-        value: String,
-    },
-
-    Legality {
-        shift: Option<u8>, // None: format absent from all loaded data — matches nothing
-        expected: u64,
-    },
-
-    ManaCostCmp {
-        op: CmpOp,
-        pips: HashMap<String, u8>,
-        cmc: f32,
-    },
-
-    Devotion {
-        op: CmpOp,
-        pips: HashMap<String, u8>,
-    },
-
-    DateCmp {
-        op: CmpOp,
-        value: u32, // yyyymmdd, partial dates zero-padded (e.g. "2026-07" → 20260700)
-    },
-
-    YearCmp {
-        op: CmpOp,
-        year: i32,
-    },
-}
-
-impl FilterExpr {
-    fn matches(&self, card: &ACard, strings: &AStrings) -> bool {
-        self.tri(card, strings) == Some(true)
-    }
-
-    /// Three-valued evaluation mirroring SQL: None is SQL's NULL ("unknown"),
-    /// produced when a compared field is missing from the card. NOT/AND/OR
-    /// propagate unknown exactly like SQL ternary logic, so -power>2 excludes
-    /// powerless cards (NOT NULL = NULL), matching Scryfall's "attribute
-    /// filters only match cards that have the attribute", while
-    /// -(power>2 and t:creature) still matches instants (NULL AND false =
-    /// false, NOT false = true). Only Some(true) counts as a match.
-    fn tri(&self, card: &ACard, strings: &AStrings) -> Option<bool> {
-        match self {
-            FilterExpr::True => Some(true),
-
-            FilterExpr::And(children) => {
-                let mut unknown = false;
-                for c in children {
-                    match c.tri(card, strings) {
-                        Some(false) => return Some(false),
-                        None => unknown = true,
-                        Some(true) => {}
-                    }
-                }
-                if unknown { None } else { Some(true) }
-            }
-            FilterExpr::Or(children) => {
-                let mut unknown = false;
-                for c in children {
-                    match c.tri(card, strings) {
-                        Some(true) => return Some(true),
-                        None => unknown = true,
-                        Some(false) => {}
-                    }
-                }
-                if unknown { None } else { Some(false) }
-            }
-            FilterExpr::Not(inner) => inner.tri(card, strings).map(|b| !b),
-
-            FilterExpr::ExactName(lower) => Some(card.card_name_lower.as_str() == lower.as_str()),
-
-            FilterExpr::NumericCmp { lhs, op, rhs } => {
-                match (lhs.eval(card), rhs.eval(card)) {
-                    (Some(a), Some(b)) => Some(cmp(*op, a, b)),
-                    _ => None, // a compared field is missing: SQL NULL
-                }
-            }
-
-            FilterExpr::TextContains { field, word } => {
-                text_search_field_value(card, strings, *field).map(|s| s.contains(word.as_str()))
-            }
-
-            FilterExpr::TextExact { field, op, value } => {
-                text_field_value(card, strings, *field).map(|s| match op {
-                    CmpOp::Eq => s == value,
-                    CmpOp::Ne => s != value,
-                    CmpOp::Lt => s < value.as_str(),
-                    CmpOp::Le => s <= value.as_str(),
-                    CmpOp::Gt => s > value.as_str(),
-                    CmpOp::Ge => s >= value.as_str(),
-                })
-            }
-
-            FilterExpr::TextRegex { field, regex } => {
-                text_field_value(card, strings, *field).map(|s| regex.is_match(s))
-            }
-
-            FilterExpr::ColorCmp { field, op, mask } => {
-                let bits = card_colors(card, *field);
-                Some(match op {
-                    CmpOp::Ge => bits & mask == *mask,
-                    CmpOp::Eq => bits == *mask,
-                    CmpOp::Le => bits & !mask == 0,
-                    CmpOp::Lt => bits & !mask == 0 && bits != *mask,
-                    CmpOp::Gt => bits & mask == *mask && bits != *mask,
-                    CmpOp::Ne => bits != *mask,
-                })
-            }
-
-            FilterExpr::TypeCmp { mask, op } => {
-                let bits = u16::from(card.card_types);
-                Some(match op {
-                    CmpOp::Ge => bits & mask != 0,
-                    CmpOp::Eq => bits == *mask,
-                    CmpOp::Le => bits & !mask == 0,
-                    CmpOp::Lt => bits & !mask == 0 && bits != *mask,
-                    CmpOp::Gt => bits & mask != 0 && bits != *mask,
-                    CmpOp::Ne => bits != *mask,
-                })
-            }
-
-            FilterExpr::CollectionCmp { field, op, value } => {
-                // Set-containment semantics against the single-value query {value},
-                // mirroring the SQL path's jsonb operators (@>, <@, =, <> and the
-                // strict variants). Lt (proper subset of a one-element set) can only
-                // be the empty collection; Ne is not-exactly-equal, NOT "lacks value"
-                // (that's what negation is for).
-                let coll = card_collection(card, *field);
-                Some(match op {
-                    CmpOp::Ge => coll.contains(value),
-                    CmpOp::Eq => coll.len() == 1 && coll.contains(value),
-                    CmpOp::Gt => coll.contains(value) && coll.len() > 1,
-                    CmpOp::Le => coll.all_equal(value),
-                    CmpOp::Lt => coll.len() == 0,
-                    CmpOp::Ne => !(coll.len() == 1 && coll.contains(value)),
-                })
-            }
-
-            FilterExpr::Legality { shift, expected } => {
-                Some(shift.is_some_and(|s| (u64::from(card.card_legalities) >> s) & 0b11 == *expected))
-            }
-
-            FilterExpr::ManaCostCmp { op, pips, cmc } => {
-                let card_cmc = f32::from(card.mana_cost.cmc);
-                let card_pips = &card.mana_cost.pips;
-                Some(match op {
-                    CmpOp::Ge => {
-                        pips.iter().all(|(sym, &n)| {
-                            card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) >= n
-                        }) && card_cmc >= *cmc
-                    }
-                    CmpOp::Le => {
-                        card_pips.iter().all(|(sym, n)| {
-                            pips.get(sym.as_str()).copied().unwrap_or(0) >= u8::from(*n)
-                        }) && card_cmc <= *cmc
-                    }
-                    CmpOp::Eq => {
-                        card_cmc == *cmc
-                            && card_pips.len() == pips.len()
-                            && pips.iter().all(|(sym, &n)| {
-                                card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) == n
-                            })
-                    }
-                    CmpOp::Gt => {
-                        let contains = pips.iter().all(|(sym, &n)| {
-                            card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) >= n
-                        }) && card_cmc >= *cmc;
-                        let exact = card_cmc == *cmc
-                            && card_pips.len() == pips.len()
-                            && pips.iter().all(|(sym, &n)| {
-                                card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) == n
-                            });
-                        contains && !exact
-                    }
-                    CmpOp::Lt => {
-                        let subset = card_pips.iter().all(|(sym, n)| {
-                            pips.get(sym.as_str()).copied().unwrap_or(0) >= u8::from(*n)
-                        }) && card_cmc <= *cmc;
-                        let exact = card_cmc == *cmc
-                            && card_pips.len() == pips.len()
-                            && pips.iter().all(|(sym, &n)| {
-                                card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) == n
-                            });
-                        subset && !exact
-                    }
-                    CmpOp::Ne => {
-                        !(card_cmc == *cmc
-                            && card_pips.len() == pips.len()
-                            && pips.iter().all(|(sym, &n)| {
-                                card_pips.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) == n
-                            }))
-                    }
-                })
-            }
-
-            FilterExpr::Devotion { op, pips } => {
-                // Mirrors the SQL path's JSONB containment on the devotion column
-                // (devotion @> query, <@, =, and the strict/negated variants):
-                // per-color positional arrays contain each other iff the counts
-                // compare, so containment reduces to count comparisons here.
-                // The card-side map can carry non-color keys (generic pips like "1"
-                // come straight from mana_cost_jsonb) that the DB's devotion column
-                // never holds, so only WUBRGC entries participate — query pips are
-                // already color-only (see build_binary).
-                let devotion = if card.mana_cost.devotion.is_some() {
-                    card.mana_cost.devotion.as_ref().unwrap()
-                } else {
-                    &card.mana_cost.pips
-                };
-                let ge = pips.iter().all(|(sym, &n)| {
-                    devotion.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) >= n
-                });
-                let le = devotion.iter()
-                    .filter(|(sym, _)| is_devotion_sym(sym.as_str()))
-                    .all(|(sym, n)| pips.get(sym.as_str()).copied().unwrap_or(0) >= u8::from(*n));
-                let eq = devotion.keys().filter(|sym| is_devotion_sym(sym.as_str())).count() == pips.len()
-                    && pips.iter().all(|(sym, &n)| {
-                        devotion.get(sym.as_str()).map(|v| u8::from(*v)).unwrap_or(0) == n
-                    });
-                Some(match op {
-                    CmpOp::Ge => ge,
-                    CmpOp::Eq => eq,
-                    CmpOp::Le => le,
-                    CmpOp::Gt => ge && !eq,
-                    CmpOp::Lt => le && !eq,
-                    CmpOp::Ne => !eq,
-                })
-            }
-
-            FilterExpr::DateCmp { op, value } => {
-                // value is a zero-padded yyyymmdd (see build_binary); zero-padding a
-                // partial date reproduces the old lexicographic-prefix semantics exactly,
-                // since any real day/month (>= 01) compares greater than 00.
-                let Some(date) = card.released_at_int.as_ref().map(|v| u32::from(*v)) else {
-                    return None; // missing date: SQL NULL
-                };
-                Some(match op {
-                    CmpOp::Eq => date == *value,
-                    CmpOp::Ne => date != *value,
-                    CmpOp::Lt => date < *value,
-                    CmpOp::Le => date <= *value,
-                    CmpOp::Gt => date > *value,
-                    CmpOp::Ge => date >= *value,
-                })
-            }
-
-            FilterExpr::YearCmp { op, year } => {
-                let Some(date) = card.released_at_int.as_ref().map(|v| u32::from(*v)) else {
-                    return None; // missing date: SQL NULL
-                };
-                let card_year = (date / 10_000) as i32;
-                Some(match op {
-                    CmpOp::Eq => card_year == *year,
-                    CmpOp::Ne => card_year != *year,
-                    CmpOp::Gt => card_year > *year,
-                    CmpOp::Lt => card_year < *year,
-                    CmpOp::Ge => card_year >= *year,
-                    CmpOp::Le => card_year <= *year,
-                })
-            }
-        }
-    }
-
-    /// Returns true if every leaf predicate touches only card-level attributes
-    /// (constant across all printings of the same oracle id). When true for a
-    /// `unique=card, prefer=default` query, the preferred-printing index covers
-    /// the full result set with no dedup needed.
-    fn is_card_level(&self) -> bool {
-        match self {
-            FilterExpr::True | FilterExpr::ExactName(_) => true,
-            FilterExpr::And(c) | FilterExpr::Or(c) => c.iter().all(|x| x.is_card_level()),
-            FilterExpr::Not(inner) => inner.is_card_level(),
-            FilterExpr::NumericCmp { lhs, rhs, .. } => lhs.is_card_level() && rhs.is_card_level(),
-            FilterExpr::TextContains { field, .. } => field.is_card_level(),
-            FilterExpr::TextExact   { field, .. } => field.is_card_level(),
-            FilterExpr::TextRegex   { field, .. } => field.is_card_level(),
-            FilterExpr::ColorCmp { .. } | FilterExpr::TypeCmp { .. } => true,
-            FilterExpr::CollectionCmp { field, .. } => field.is_card_level(),
-            FilterExpr::Legality { .. } | FilterExpr::ManaCostCmp { .. } | FilterExpr::Devotion { .. } => true,
-            FilterExpr::DateCmp { .. } | FilterExpr::YearCmp { .. } => false,
-        }
-    }
-}
-
-// ─── Building FilterExpr from JSON ───────────────────────────────────────────
-
-fn str_op_to_cmp(s: &str) -> Result<CmpOp, String> {
-    match s {
-        "=" | ":" => Ok(CmpOp::Eq),
-        "!="      => Ok(CmpOp::Ne),
-        "<"       => Ok(CmpOp::Lt),
-        "<="      => Ok(CmpOp::Le),
-        ">"       => Ok(CmpOp::Gt),
-        ">="      => Ok(CmpOp::Ge),
-        _ => Err(format!("unknown operator: {s}")),
-    }
-}
-
-fn op_to_collection_cmp(op: &str) -> CmpOp {
-    match op {
-        ":" | ">=" => CmpOp::Ge,
-        "="        => CmpOp::Eq,
-        ">"        => CmpOp::Gt,
-        "<="       => CmpOp::Le,
-        "<"        => CmpOp::Lt,
-        "!="       => CmpOp::Ne,
-        _          => CmpOp::Ge,
-    }
-}
-
-fn op_to_color_cmp(op: &str) -> CmpOp {
-    match op {
-        ":" | ">=" => CmpOp::Ge,
-        "="        => CmpOp::Eq,
-        "<="       => CmpOp::Le,
-        "<"        => CmpOp::Lt,
-        ">"        => CmpOp::Gt,
-        "!="       => CmpOp::Ne,
-        _          => CmpOp::Ge,
-    }
-}
-
-fn build_num_expr(v: &Value) -> Result<NumExpr, String> {
-    let node_type = v["node_type"].as_str().unwrap_or("");
-    let kw = &v["kwargs"];
-    match node_type {
-        "NumericValueNode" => {
-            let val = kw["value"].as_f64().ok_or("NumericValueNode missing value")?;
-            Ok(NumExpr::Const(val))
-        }
-        "CardAttributeNode" => {
-            let attr = kw["attribute_name"].as_str().unwrap_or("");
-            attr_to_num_field(attr)
-                .map(NumExpr::Field)
-                .ok_or_else(|| format!("unknown numeric field: {attr}"))
-        }
-        "CardBinaryOperatorNode" => {
-            let op_str = kw["op"].as_str().unwrap_or("");
-            let arith_op = match op_str {
-                "+" => ArithOp::Add,
-                "-" => ArithOp::Sub,
-                "*" => ArithOp::Mul,
-                "/" => ArithOp::Div,
-                _ => return Err(format!("expected arithmetic op, got: {op_str}")),
-            };
-            let lhs = build_num_expr(&kw["lhs"])?;
-            let rhs = build_num_expr(&kw["rhs"])?;
-            Ok(NumExpr::Arith(Box::new(lhs), arith_op, Box::new(rhs)))
-        }
-        _ => Err(format!("unexpected node in numeric expr: {node_type}")),
-    }
-}
-
-fn build_filter(v: &Value) -> Result<FilterExpr, String> {
-    let node_type = v["node_type"].as_str().unwrap_or("");
-    let kw = &v["kwargs"];
-
-    match node_type {
-        "TrueNode" => Ok(FilterExpr::True),
-
-        "AndNode" => {
-            let operands = kw["operands"]
-                .as_array()
-                .ok_or("AndNode missing operands")?
-                .iter()
-                .map(build_filter)
-                .collect::<Result<Vec<_>, _>>()?;
-            Ok(FilterExpr::And(operands))
-        }
-
-        "OrNode" => {
-            let operands = kw["operands"]
-                .as_array()
-                .ok_or("OrNode missing operands")?
-                .iter()
-                .map(build_filter)
-                .collect::<Result<Vec<_>, _>>()?;
-            Ok(FilterExpr::Or(operands))
-        }
-
-        "NotNode" => {
-            let inner = build_filter(&kw["operand"])?;
-            Ok(FilterExpr::Not(Box::new(inner)))
-        }
-
-        "ExactNameNode" => {
-            let value = kw["value"].as_str().unwrap_or("").to_string();
-            Ok(FilterExpr::ExactName(value))
-        }
-
-        "CardBinaryOperatorNode" => build_binary(kw),
-
-        _ => Err(format!("unexpected top-level node type: {node_type}")),
-    }
-}
-
-fn build_binary(kw: &Value) -> Result<FilterExpr, String> {
-    let op = kw["op"].as_str().unwrap_or(":");
-    let lhs = &kw["lhs"];
-    let rhs = &kw["rhs"];
-
-    let lhs_type = lhs["node_type"].as_str().unwrap_or("");
-    let lhs_kw   = &lhs["kwargs"];
-
-    if lhs_type != "CardAttributeNode" {
-        let lhs_expr = build_num_expr(lhs)?;
-        let rhs_expr = build_num_expr(rhs)?;
-        let cmp_op   = str_op_to_cmp(op)?;
-        return Ok(FilterExpr::NumericCmp { lhs: lhs_expr, op: cmp_op, rhs: rhs_expr });
-    }
-
-    let attr = lhs_kw["attribute_name"].as_str().unwrap_or("");
-    let orig = lhs_kw["original_attribute"].as_str().unwrap_or("");
-
-    if let Some(num_field) = attr_to_num_field(attr) {
-        let cmp_op   = str_op_to_cmp(op)?;
-        let rhs_expr = build_num_expr(rhs)?;
-        return Ok(FilterExpr::NumericCmp { lhs: NumExpr::Field(num_field), op: cmp_op, rhs: rhs_expr });
-    }
-
-    if attr == "released_at" {
-        let val_str = rhs_value_str(rhs);
-        if orig == "year" {
-            let year: i32 = val_str.parse().map_err(|_| format!("bad year: {val_str}"))?;
-            let cmp_op = str_op_to_cmp(op)?;
-            return Ok(FilterExpr::YearCmp { op: cmp_op, year });
-        }
-        let cmp_op = str_op_to_cmp(op)?;
-        // yyyymmdd as integer; zero-pad partial dates so ordering matches the
-        // lexicographic compare on ISO strings this replaced (day 00 < any real day).
-        let digits: String = val_str.chars().filter(|c| c.is_ascii_digit()).collect();
-        if digits.is_empty() || digits.len() > 8 {
-            return Err(format!("bad date: {val_str}"));
-        }
-        let value: u32 = format!("{digits:0<8}").parse().map_err(|_| format!("bad date: {val_str}"))?;
-        return Ok(FilterExpr::DateCmp { op: cmp_op, value });
-    }
-
-    if attr == "mana_cost_jsonb" {
-        let mana_str = rhs_value_str(rhs);
-        let pips = mana_pip_counts(mana_str);
-        let cmc  = mana_cmc(mana_str);
-        let cmp_op = match op { ":" => CmpOp::Ge, _ => str_op_to_cmp(op)? };
-        return Ok(FilterExpr::ManaCostCmp { op: cmp_op, pips, cmc });
-    }
-
-    if attr == "devotion" {
-        let mana_str = rhs_value_str(rhs);
-        // Split hybrid symbols ({R/G} -> R:1, G:1) and keep only WUBRGC, matching
-        // calculate_devotion() in SQL (which counts only color characters).
-        // mana_pip_counts is NOT used directly because it keeps hybrids as single keys.
-        let mut pips: HashMap<String, u8> = HashMap::new();
-        for (sym, n) in mana_pip_counts(mana_str) {
-            if sym.contains('/') {
-                for part in sym.split('/') {
-                    if is_devotion_sym(part) {
-                        *pips.entry(part.to_string()).or_insert(0) += n;
-                    }
-                }
-            } else if is_devotion_sym(&sym) {
-                *pips.entry(sym).or_insert(0) += n;
-            }
-        }
-        let cmp_op = match op { ":" => CmpOp::Ge, _ => str_op_to_cmp(op)? };
-        return Ok(FilterExpr::Devotion { op: cmp_op, pips });
-    }
-
-    if matches!(attr, "card_colors" | "card_color_identity" | "produced_mana") {
-        let color_field = match attr {
-            "card_colors"          => ColorField::Colors,
-            "card_color_identity"  => ColorField::ColorIdentity,
-            _                      => ColorField::ProducedMana,
-        };
-        let color_strs: Vec<&str> = rhs
-            .as_array()
-            .map(|a| a.iter().filter_map(|v| v.as_str()).collect())
-            .unwrap_or_default();
-        let mask = color_list_to_mask(&color_strs);
-        // id:/identity: means "card's identity is a subset of query colors" (Le), not superset (Ge)
-        let cmp_op = if attr == "card_color_identity" && op == ":" {
-            CmpOp::Le
-        } else {
-            op_to_color_cmp(op)
-        };
-        return Ok(FilterExpr::ColorCmp { field: color_field, op: cmp_op, mask });
-    }
-
-    if attr == "card_legalities" {
-        let format = rhs
-            .as_array()
-            .and_then(|a| a.first())
-            .and_then(|v| v.as_str())
-            .unwrap_or("");
-        let expected = match orig {
-            "format" | "f" | "legal" => LEGALITY_LEGAL,
-            "banned"                 => LEGALITY_BANNED,
-            "restricted"             => LEGALITY_RESTRICTED,
-            _                        => LEGALITY_LEGAL,
-        };
-        return Ok(FilterExpr::Legality { shift: format_shift(format), expected });
-    }
-
-    if attr == "card_types" {
-        let mask: u16 = rhs
-            .as_array()
-            .map(|a| a.iter().fold(0u16, |acc, v| acc | card_type_str_to_bit(v.as_str().unwrap_or(""))))
-            .unwrap_or(0);
-        return Ok(FilterExpr::TypeCmp { mask, op: op_to_collection_cmp(op) });
-    }
-
-    if attr == "card_subtypes" {
-        let value = rhs.as_array().and_then(|a| a.first()).and_then(|v| v.as_str()).unwrap_or("").to_string();
-        return Ok(FilterExpr::CollectionCmp { field: CollField::Subtypes, op: op_to_collection_cmp(op), value });
-    }
-
-    if attr == "card_keywords" {
-        let value  = rhs.as_array().and_then(|a| a.first()).and_then(|v| v.as_str()).unwrap_or("").to_string();
-        let cmp_op = op_to_collection_cmp(op);
-        return Ok(FilterExpr::CollectionCmp { field: CollField::Keywords, op: cmp_op, value });
-    }
-
-    if matches!(attr, "card_oracle_tags" | "card_art_tags" | "card_is_tags" | "card_frame_data") {
-        let coll_field = match attr {
-            "card_oracle_tags" => CollField::OracleTags,
-            "card_art_tags"    => CollField::ArtTags,
-            "card_is_tags"     => CollField::IsTags,
-            _                  => CollField::FrameData,
-        };
-        let value  = rhs.as_array().and_then(|a| a.first()).and_then(|v| v.as_str()).unwrap_or("").to_string();
-        let cmp_op = op_to_collection_cmp(op);
-        return Ok(FilterExpr::CollectionCmp { field: coll_field, op: cmp_op, value });
-    }
-
-    build_text_filter(attr, op, rhs)
-}
-
-fn rhs_value_str<'a>(rhs: &'a Value) -> &'a str {
-    rhs["kwargs"]["value"].as_str().unwrap_or("")
-}
-
-fn build_text_filter(attr: &str, op: &str, rhs: &Value) -> Result<FilterExpr, String> {
-    let rhs_node_type = rhs["node_type"].as_str().unwrap_or("");
-
-    if rhs_node_type == "RegexValueNode" {
-        let pattern  = rhs["kwargs"]["value"].as_str().unwrap_or("");
-        let re = Regex::new(&format!("(?i){pattern}"))
-            .map_err(|e| format!("invalid regex '{pattern}': {e}"))?;
-        let field = match attr {
-            "card_name"   => TextField::NameLower,
-            "oracle_text" => TextField::OracleTextLower,
-            "flavor_text" => TextField::FlavorTextLower,
-            "card_artist" => TextField::ArtistLower,
-            _ => return Err(format!("regex not supported on {attr}")),
-        };
-        return Ok(FilterExpr::TextRegex { field, regex: re });
-    }
-
-    let raw_value = rhs["kwargs"]["value"].as_str().unwrap_or("");
-
-    if matches!(attr, "card_set_code" | "card_layout" | "card_border" | "card_watermark" | "collector_number") {
-        // collector_number_id is stored raw and mixed-case (e.g. "10E-105"); compare exactly,
-        // matching the SQL path. The other four are lowercased at import, so lowercasing
-        // the query value gives case-insensitive matching with a plain equality.
-        let value = if attr == "collector_number" { raw_value.to_string() } else { raw_value.to_lowercase() };
-        let cmp_op = str_op_to_cmp(op)?;
-        let field = match attr {
-            "card_set_code"    => TextField::SetCode,
-            "card_layout"      => TextField::Layout,
-            "card_border"      => TextField::Border,
-            "card_watermark"   => TextField::Watermark,
-            "collector_number" => TextField::CollectorNumber,
-            _                  => unreachable!(),
-        };
-        return Ok(FilterExpr::TextExact { field, op: cmp_op, value });
-    }
-
-    let lower_word = raw_value.to_lowercase();
-    if op == ":" {
-        let tsf = match attr {
-            "card_name"   => TextSearchField::NameLower,
-            "oracle_text" => TextSearchField::OracleTextLower,
-            "flavor_text" => TextSearchField::FlavorTextLower,
-            "card_artist" => TextSearchField::ArtistLower,
-            _ => return Err(format!("text substring not supported on {attr}")),
-        };
-        return Ok(FilterExpr::TextContains { field: tsf, word: lower_word });
-    }
-
-    let field = match attr {
-        "card_name"   => TextField::NameLower,
-        "oracle_text" => TextField::OracleTextLower,
-        "flavor_text" => TextField::FlavorTextLower,
-        "card_artist" => TextField::ArtistLower,
-        _ => return Err(format!("unknown text field: {attr}")),
-    };
-    let cmp_op = str_op_to_cmp(op)?;
-    Ok(FilterExpr::TextExact { field, op: cmp_op, value: raw_value.to_lowercase() })
-}
+mod filter;
+use filter::*;
 
 // ─── Trigram index ────────────────────────────────────────────────────────────
 
@@ -1907,20 +810,6 @@ struct CardData {
     // it is shared across all workers at zero per-worker heap cost.
     // ~31k entries × 4 bytes = ~124 KB in the archive.
     preferred_indices: Vec<u32>,
-}
-
-/// Adopt the archive's format→shift assignments into this process's registry.
-/// Cheap no-op (one read lock) once the registry has caught up.
-fn sync_format_shifts(archived: &Archived<HashMap<String, u8>>) {
-    let behind = format_shifts().read().map(|m| m.len() < archived.len()).unwrap_or(false);
-    if !behind {
-        return;
-    }
-    if let Ok(mut shifts) = format_shifts().write() {
-        for (format, shift) in archived.iter() {
-            shifts.insert(format.as_str().to_string(), *shift);
-        }
-    }
 }
 
 // ─── Candidate narrowing ─────────────────────────────────────────────────────
@@ -2875,271 +1764,4 @@ mod card_engine {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use rkyv::rancor::Error;
-
-    /// Build a TrigramIndex mapping each word's trigrams to the given card ids.
-    fn index_of(words: &[(&str, &[u32])]) -> TrigramIndex {
-        let mut idx: TrigramIndex = HashMap::new();
-        for (word, cards) in words {
-            for w in word.as_bytes().windows(3) {
-                let entry = idx.entry([w[0], w[1], w[2]]).or_default();
-                for &c in *cards {
-                    if !entry.contains(&c) { entry.push(c); }
-                }
-                entry.sort_unstable();
-            }
-        }
-        idx
-    }
-
-    /// Archive the index and query it, matching how the engine reads the shared snapshot.
-    fn candidates(idx: &TrigramIndex, word: &str) -> Option<Vec<u32>> {
-        let bytes = rkyv::to_bytes::<Error>(idx).expect("serialize trigram index");
-        let archived = rkyv::access::<Archived<TrigramIndex>, Error>(&bytes).expect("access trigram index");
-        trigram_candidates(archived, word)
-    }
-
-    #[test]
-    fn trigram_short_word_cannot_narrow() {
-        let idx = index_of(&[("bolt", &[1, 2])]);
-        assert_eq!(candidates(&idx, "bo"), None);
-    }
-
-    #[test]
-    fn trigram_all_present_intersects() {
-        // "bol" → {1,2,3}, "olt" → {1,2}: intersection is {1,2}
-        let idx = index_of(&[("bol", &[1, 2, 3]), ("olt", &[1, 2])]);
-        assert_eq!(candidates(&idx, "bolt"), Some(vec![1, 2]));
-    }
-
-    #[test]
-    fn trigram_missing_means_no_candidates() {
-        // "bol" is indexed but "olx" appears in no card, which proves nothing can
-        // match — the result must be the empty candidate set, not an intersection
-        // of whichever trigrams happen to exist.
-        let idx = index_of(&[("bolt", &[1, 2])]);
-        assert_eq!(candidates(&idx, "bolx"), Some(Vec::new()));
-    }
-
-    #[test]
-    fn trigram_fully_unindexed_word_is_empty_not_unnarrowed() {
-        let idx = index_of(&[("bolt", &[1, 2])]);
-        assert_eq!(candidates(&idx, "zzz"), Some(Vec::new()));
-    }
-
-    // Verify that HashMap<[u8; 3], Vec<u32>> (the trigram index key type) round-trips
-    // through rkyv and supports lookup via the same [u8; 3] key type.
-    #[test]
-    fn test_trigram_index_archive_and_lookup() {
-        let mut idx: HashMap<[u8; 3], Vec<u32>> = HashMap::new();
-        idx.insert([b'a', b'b', b'c'], vec![1, 2, 3]);
-        idx.insert([b'x', b'y', b'z'], vec![4, 5]);
-        idx.insert([b'f', b'o', b'o'], vec![10, 20, 30, 40]);
-
-        let bytes = rkyv::to_bytes::<Error>(&idx).expect("serialize trigram index");
-        let archived = rkyv::access::<rkyv::Archived<HashMap<[u8; 3], Vec<u32>>>, Error>(&bytes)
-            .expect("access trigram index");
-
-        // Key present -- length check
-        let abc = archived.get(&[b'a', b'b', b'c']).expect("abc must be present");
-        assert_eq!(abc.len(), 3);
-
-        // Key present -- value iteration (elements are rend::u32_le, not u32)
-        let foo: Vec<u32> = archived
-            .get(&[b'f', b'o', b'o'])
-            .expect("foo must be present")
-            .iter()
-            .map(|x| u32::from(*x))
-            .collect();
-        assert_eq!(foo, vec![10, 20, 30, 40]);
-
-        // Key absent
-        assert!(archived.get(&[b'z', b'z', b'z']).is_none());
-    }
-
-    // Verify that HashSet<String> supports contains() via &str (Borrow-based lookup).
-    // This is used for card_keywords, card_oracle_tags, card_is_tags, card_frame_data.
-    #[test]
-    fn test_hashset_string_str_lookup() {
-        let mut set: HashSet<String> = HashSet::new();
-        set.insert("Flying".to_string());
-        set.insert("Vigilance".to_string());
-        set.insert("Trample".to_string());
-
-        let bytes = rkyv::to_bytes::<Error>(&set).expect("serialize hashset");
-        let archived = rkyv::access::<rkyv::Archived<HashSet<String>>, Error>(&bytes)
-            .expect("access hashset");
-
-        assert!(archived.contains("Flying"));
-        assert!(archived.contains("Trample"));
-        assert!(!archived.contains("Deathtouch"));
-    }
-
-    // Verify that HashMap<String, Vec<u32>> (the tag index value type) supports
-    // get() via &str, which is needed for narrow_candidates tag lookups.
-    #[test]
-    fn test_tag_index_str_lookup() {
-        let mut idx: HashMap<String, Vec<u32>> = HashMap::new();
-        idx.insert("merfolk".to_string(), vec![1, 5, 9]);
-        idx.insert("dragon".to_string(), vec![2, 7]);
-
-        let bytes = rkyv::to_bytes::<Error>(&idx).expect("serialize tag index");
-        let archived = rkyv::access::<rkyv::Archived<HashMap<String, Vec<u32>>>, Error>(&bytes)
-            .expect("access tag index");
-
-        let merfolk: Vec<u32> = archived
-            .get("merfolk")
-            .expect("merfolk must be present")
-            .iter()
-            .map(|x| u32::from(*x))
-            .collect();
-        assert_eq!(merfolk, vec![1, 5, 9]);
-        assert!(archived.get("angel").is_none());
-    }
-
-    /// Measures checked rkyv::access vs access_unchecked on a production-scale
-    /// archive. This is the evidence behind the access_unchecked safety comments
-    /// on query()/size(): checked access re-validates the entire archive graph
-    /// on every call, which is milliseconds per call at ~100k cards — orders of
-    /// magnitude over total query time. Run with:
-    ///   cargo test --release -- --ignored bench_checked_vs_unchecked --nocapture
-    #[test]
-    #[ignore]
-    fn bench_checked_vs_unchecked_access() {
-        const N: usize = 96_000;
-        let words = ["draw", "card", "creature", "destroy", "target", "flying", "counter", "spell", "token", "exile"];
-        let mut interner = Interner::new();
-        let mut cards: Vec<Card> = Vec::with_capacity(N);
-        for i in 0..N {
-            let name = format!("Benchmark Card Number {i}");
-            // Oracle text keyed on i/3 so printings share texts ~3x, matching the
-            // real dataset's duplication (and exercising the CSR oracle index).
-            let group = i / 3;
-            let oracle = format!(
-                "{}: {} a {} {}, then {} {} cards. This text is representative filler standing in for \
-                 real oracle text so string validation cost is realistic for card group {group}.",
-                words[group % 10], words[(group + 1) % 10], words[(group + 2) % 10],
-                words[(group + 3) % 10], words[(group + 4) % 10], words[(group + 5) % 10],
-            );
-            let flavor = format!("Flavor text for card {i}, roughly the length of a real flavor quote in the dataset.");
-            cards.push(Card {
-                card_name_lower: InlineStr::from_str(&name.to_lowercase()),
-                card_colors: (i % 32) as u8,
-                card_color_identity: (i % 32) as u8,
-                produced_mana: 0,
-                card_types: TYPE_CREATURE,
-                scryfall_id: (i + 1) as u128,
-                oracle_id: (group + 1) as u128,
-                illustration_id: (i + 1) as u128,
-                card_name_id: interner.intern(name.clone()),
-                oracle_text_id: interner.intern(oracle.clone()),
-                oracle_text_lower_id: interner.intern(oracle.to_lowercase()),
-                flavor_text_id: interner.intern(flavor.clone()),
-                flavor_text_lower_id: interner.intern(flavor.to_lowercase()),
-                card_artist_id: interner.intern(format!("Artist {}", i % 1000)),
-                card_artist_lower_id: interner.intern(format!("artist {}", i % 1000)),
-                card_set_code: InlineStr::from_str("bench"),
-                card_layout_id: interner.intern("normal".to_string()),
-                card_border_id: interner.intern("black".to_string()),
-                card_watermark_id: NONE_STR,
-                collector_number_id: interner.intern(format!("{}", i % 500)),
-                mana_cost_text_id: interner.intern("{2}{G}{G}".to_string()),
-                type_line_id: interner.intern("Creature — Benchmark".to_string()),
-                set_name_id: interner.intern(format!("Benchmark Set {}", i % 300)),
-                released_at_int: Some(20240101),
-                cmc: Some((i % 8) as u8),
-                creature_power: Some((i % 10) as i8),
-                creature_toughness: Some((i % 10) as i8),
-                planeswalker_loyalty: None,
-                card_rarity_int: Some((i % 4) as u8),
-                collector_number_int: Some((i % 500) as u16),
-                edhrec_rank: Some(i as u32),
-                price_usd: Some(1.0),
-                price_eur: Some(1.0),
-                price_tix: Some(0.1),
-                prefer_score: None,
-                cubecobra_score: None,
-                card_subtypes: vec!["Benchmark".to_string(), words[i % 10].to_string()],
-                card_keywords: HashSet::from([words[i % 10].to_string()]),
-                card_legalities: 0,
-                card_oracle_tags: HashSet::from([format!("tag-{}", i % 100)]),
-                card_art_tags: HashSet::new(),
-                card_is_tags: HashSet::new(),
-                card_frame_data: HashSet::new(),
-                mana_cost: ManaCost {
-                    pips: HashMap::from([("G".to_string(), 2u8)]),
-                    devotion: None,
-                    cmc: (i % 8) as f32,
-                },
-                creature_power_text_id: NONE_STR,
-                creature_toughness_text_id: NONE_STR,
-            });
-        }
-        let strings = interner.strings;
-
-        let indexes = CardIndexes {
-            name_trigram:   build_trigram_index(&cards, |c| c.card_name_lower.as_str()),
-            oracle_trigram: build_oracle_text_index(&cards, &strings),
-            cmc:            build_numeric_index(&cards, |c| c.cmc.map(|v| v as i16)),
-            power:          build_numeric_index(&cards, |c| c.creature_power.map(|v| v as i16)),
-            toughness:      build_numeric_index(&cards, |c| c.creature_toughness.map(|v| v as i16)),
-            type_bits:      build_type_index(&cards),
-            subtypes:       build_list_index(&cards, |c| &c.card_subtypes),
-            keywords:       build_tag_index(&cards, |c| &c.card_keywords),
-            oracle_tags:    build_tag_index(&cards, |c| &c.card_oracle_tags),
-            art_tags:       build_tag_index(&cards, |c| &c.card_art_tags),
-            is_tags:        build_tag_index(&cards, |c| &c.card_is_tags),
-        };
-        let data = CardData { cards, strings, indexes, format_shifts: HashMap::new() };
-        let bytes = rkyv::to_bytes::<Error>(&data).expect("serialize");
-        println!("archive size: {:.1} MB", bytes.len() as f64 / 1e6);
-
-        const ITERS: u32 = 10;
-        let t = std::time::Instant::now();
-        for _ in 0..ITERS {
-            let a = rkyv::access::<Archived<CardData>, Error>(&bytes).expect("checked access");
-            assert_eq!(a.cards.len(), N);
-        }
-        let checked = t.elapsed() / ITERS;
-
-        let t = std::time::Instant::now();
-        for _ in 0..ITERS {
-            let a = unsafe { rkyv::access_unchecked::<Archived<CardData>>(&bytes) };
-            assert_eq!(a.cards.len(), N);
-        }
-        let unchecked = t.elapsed() / ITERS;
-
-        println!("checked rkyv::access:   {checked:?} per call");
-        println!("access_unchecked:       {unchecked:?} per call");
-    }
-
-    // Verify that narrow_candidates returns the correct card ids for an art tag
-    // that is present in the index, and returns None (no narrowing) for an absent tag.
-    #[test]
-    fn narrow_candidates_art_tags() {
-        let mut art_tags: TagIndex = HashMap::new();
-        art_tags.insert("wolf".to_string(), vec![0, 2]);
-        art_tags.insert("dragon".to_string(), vec![1]);
-
-        let indexes = CardIndexes { art_tags, ..Default::default() };
-        let bytes = rkyv::to_bytes::<Error>(&indexes).expect("serialize");
-        let archived = rkyv::access::<Archived<CardIndexes>, Error>(&bytes).expect("access");
-
-        let present = FilterExpr::CollectionCmp {
-            field: CollField::ArtTags,
-            op: CmpOp::Ge,
-            value: "wolf".to_string(),
-        };
-        assert_eq!(narrow_candidates(&present, archived), Some(vec![0, 2]));
-
-        // A tag not in the index cannot narrow; the eval step handles correctness.
-        let absent = FilterExpr::CollectionCmp {
-            field: CollField::ArtTags,
-            op: CmpOp::Ge,
-            value: "zombie".to_string(),
-        };
-        assert_eq!(narrow_candidates(&absent, archived), None);
-    }
-}
+mod tests;
